requirements.txt
flet==0.24.1
sqlalchemy==2.0.48
psycopg2-binary>=2.9.9
qrcode==7.4.2
pillow>=10.4.0
nodriver>=0.35
selenium>=4.0.0
undetected-chromedriver>=3.5.0
requests>=2.28.0
psutil>=5.9.0
pyinstaller>=6.0.0
pystray>=0.19.5
